fix: `Storage` indexes `int` -> `size_t`

+++ storage/storage.h
@@ -40,7 +40,8 @@ public:
         return m_size;
     }
 
-    El& operator[](int pos) {
+    El& operator[](size_t pos) {
+
         if (pos < m_size && pos >= 0) {
             return m_storage[pos];
         }
@@ -90,14 +91,14 @@ template<typename El>
 void sortStorage(Storage<El>& st);
 
 template<typename T>
-int searchInsertPos(const Storage<T>& arr, const T& value) {
+size_t searchInsertPos(const Storage<T>& arr, const T& value) {
 
-    int left = 0;
-    int right = arr.size() - 1;
+    size_t left = 0;
+    size_t right = arr.size() - 1;
 
     while (left <= right) {
 
-        int mid = left + (right - left) / 2;
+        size_t mid = left + (right - left) / 2;
 
         if (arr[mid] == value) {
             return mid;
